scripts/requirements.txt
absl-py==2.4.0
appnope==0.1.4
asttokens==3.0.1
astunparse==1.6.3
certifi==2026.4.22
charset-normalizer==3.4.7
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.16.0
idna==3.13
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.14.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pandas==3.0.2
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
protobuf==4.25.9
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.33.1
rich==15.0.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.3.0
threadpoolctl==3.6.0
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.8
wrapt==2.1.2
